Move deprecation helpers into separate module
