+++ packages/liveblocks-client/src/client.ts
@@ -1,3 +1,4 @@
+import { deprecateIf } from "./deprecation";
 import type { InternalRoom } from "./room";
 import { createRoom } from "./room";
 import type {
@@ -9,7 +10,6 @@ import type {
   Room,
   RoomInitializers,
 } from "./types";
-import { deprecateIf } from "./utils";
 
 type EnterOptions<TPresence, TStorage> = Resolve<
   // Enter options are just room initializers, plus an internal option
