+++ packages/liveblocks-client/src/internal.ts
@@ -19,7 +19,7 @@ export type {
 export type { Resolve, RoomInitializers } from "./types";
 
 export { ClientMessageType, CrdtType, OpType, ServerMessageType } from "./live";
-export { deprecate, deprecateIf } from "./utils";
+export { deprecate, deprecateIf } from "./deprecation";
 
 export {
   lsonToJson,
