+++ packages/liveblocks-client/src/utils.ts
@@ -23,41 +23,6 @@ import type {
   StorageUpdate,
 } from "./types";
 
-// Keeps a set of deprecation messages in memory that it has warned about
-// already. There will be only one deprecation message in the console, no
-// matter how often it gets called.
-const _emittedDeprecationWarnings: Set<string> = new Set();
-
-/**
- * Displays a deprecation warning in the dev console. Only in dev mode, and
- * only once per message/key. In production, this is a no-op.
- */
-export function deprecate(message: string, key = message) {
-  if (process.env.NODE_ENV !== "production") {
-    if (!_emittedDeprecationWarnings.has(key)) {
-      _emittedDeprecationWarnings.add(key);
-      console.error(`DEPRECATION WARNING: ${message}`);
-    }
-  }
-}
-
-/**
- * Conditionally displays a deprecation warning in the dev
- * console if the first argument is truthy. Only in dev mode, and
- * only once per message/key. In production, this is a no-op.
- */
-export function deprecateIf(
-  condition: unknown,
-  message: string,
-  key = message
-) {
-  if (process.env.NODE_ENV !== "production") {
-    if (condition) {
-      deprecate(message, key);
-    }
-  }
-}
-
 export function remove<T>(array: T[], item: T) {
   for (let i = 0; i < array.length; i++) {
     if (array[i] === item) {
